erp-multisede/requirements/dev.txt
Django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.*
django-debug-toolbar==4.*
django-environ==0.11.*
psycopg[binary,pool]==3.2.*
celery==5.4.*
redis==5.0.*
drf-spectacular==0.27.*
django-filter==24.*
django-axes==6.*
django-guardian==2.*
django-simple-history==3.*
pytest
pytest-django
pytest-cov
